superdupertax/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.3
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.7.1
excel-base==1.0.4
inflection==0.5.1
isoweek==1.3.3
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
xlwt==1.3.0
